# Conflicts:
#	Auction_frontend/src/app/core/components/landing-page/landing-page.component.html

+++ Auction_backend/src/main/java/com/online/auction/model/Account.java
@@ -15,7 +15,7 @@ public class Account {
     @Id
     private int userId;
 
-    private float funds;
+    private double funds;
 
     @OneToOne
     @JoinColumn(name = "user_id")
